feat(empty-states): work in progress part one
