+++ packages/blockchain-wallet-v4-frontend/src/components/Layout/index.ts
@@ -1,3 +1,5 @@
+import { Box } from '../Box'
+import { Text } from 'blockchain-info-components'
 import styled from 'styled-components'
 
 export const StickyHeader = styled.div`
@@ -22,4 +24,50 @@ export const SceneWrapper = styled.div<{ centerContent?: boolean }>`
     align-items: center;
     justify-content: center;
   `}
+`
+
+// Header, Icon, Subheaders for all empty states
+
+export const IconBackground = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 40px;
+  height: 40px;
+  min-width: 40px;
+  background-color: ${props => props.theme.blue000};
+  border-radius: 40px;
+`
+export const SceneHeader = styled.div`
+  display: flex;
+  flex-direction: row;
+  margin-bottom: 20px;
+  > :first-child {
+    margin-right: 14px;
+  }
+`
+export const SceneHeaderText = styled(Text)`
+  font-size: 32px;
+  color: black;
+  font-weight: 600;
+`
+export const SceneSubHeaderText = styled(Text)`
+  font-size: 16px;
+  color: ${props => props.theme.grey600};
+  font-weight: 500;
+  margin-bottom: 30px;
+`
+
+export const CustomBox = styled(Box)`
+  background-image: url('/img/buy-sell-learn-more.png');
+  /* stylelint-disable */
+  background-image: -webkit-image-set(
+    url('/img/buy-sell-learn-more.png') 1x,
+    url('/img/buy-sell-learn-more@2x.png') 2x
+  );
+  /* stylelint-enable */
+  background-repeat: no-repeat;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
 `
