+++ packages/blockchain-wallet-v4-frontend/src/scenes/TheExchange/template.info.js
@@ -1,0 +1,11 @@
+import { CustomBox } from 'components/Layout'
+
+import React from 'react'
+
+class ExchangeInfo extends React.PureComponent {
+  render () {
+    return <CustomBox />
+  }
+}
+
+export default ExchangeInfo
